// OldCode/src/interfaces/orderObject.ts
export interface OrderObject {    
    SuppressPrinting: boolean;
    ResourceType: string;
    ObjectType: number;
    Text: string;
    ObjectName: string;
    FontStyle: string;
    X: number;
    Y: number;
    Width: number;
    Height: number;
    FontSize: number;
    OrgFontSize: number;
    Font: string;
    TextAlign: string;
    FontColor: string;
    PageNumber: number;
    ObjectGroup: string;
    FlagResize: boolean;
    FlagWordBreak: boolean;
    FlagVisible: boolean;
    DisplayOrder: number;
    zIndex: number;
    Alpha: number;
    Angle: number;
    Radius: number;
    ID: number;
    CouponID: number;
    ProductID: number;
    LinkedObject: number;
    ImageAlign: string;
    BackgroundColor: string;
}